refactor tests into unit and versioned types
